refactor: move OutlineHeading into SharedComponents
