+++ src/components/BubbleMessage/BubbleMessage.tsx
@@ -1,8 +1,8 @@
 import { FC } from "react";
 
-import OutlineHeading from "../OutlineHeading";
 import { OakColorName } from "../../styles/theme/types";
 
+import OutlineHeading from "@/components/SharedComponents/OutlineHeading";
 import Svg from "@/components/SharedComponents/Svg";
 import { Heading } from "@/components/SharedComponents/Typography";
 import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
