+++ src/components/pages/CurriculumInfo/tabs/UnitsTab/UnitsTab.tsx
@@ -6,7 +6,7 @@ import Flex from "@/components/SharedComponents/Flex";
 import Card from "@/components/SharedComponents/Card/Card";
 import { CurriculumUnitsTabData } from "@/node-lib/curriculum-api-2023";
 import Icon from "@/components/SharedComponents/Icon";
-import OutlineHeading from "@/components/OutlineHeading/OutlineHeading";
+import OutlineHeading from "@/components/SharedComponents/OutlineHeading/OutlineHeading";
 import Button from "@/components/SharedComponents/Button/Button";
 import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders/BrushBorders";
 import Grid, { GridArea } from "@/components/SharedComponents/Grid";
